pyproject.toml
[project]
name = "coding-assignment-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.2",
    "smolagents[litellm]>=1.2.2",
]
